4sprint

+++ search-server/main.cpp
@@ -1,18 +1,19 @@
-#include <iostream> 
-#include <string>  
-#include <vector> 
+#include <iostream>
+#include <string>
+#include <vector>
 #include <set>
 #include <map>
 
-#include "search_server.h"
 #include "paginator.h"
 #include "document.h"
 #include "request_queue.h"
+#include "search_server.h"
 
 using namespace std;
 
 template <typename First, typename Second>
-std::ostream& operator << (std::ostream& out, const std::pair<First, Second>& container) {
+std::ostream &operator<<(std::ostream &out, const std::pair<First, Second> &container)
+{
     out << "(";
     out << container.first << ", " << container.second;
     out << ")";
@@ -20,10 +21,13 @@ std::ostream& operator << (std::ostream& out, const std::pair<First, Second>& co
 }
 
 template <typename Container>
-void Print(std::ostream& out, const Container& container) {
+void Print(std::ostream &out, const Container &container)
+{
     bool flag = true;
-    for (const auto& element : container) {
-        if (!flag) {
+    for (const auto &element : container)
+    {
+        if (!flag)
+        {
             out << ", ";
         }
 
@@ -33,7 +37,8 @@ void Print(std::ostream& out, const Container& container) {
 }
 
 template <typename Term>
-std::ostream& operator << (std::ostream& out, const std::vector<Term>& container) {
+std::ostream &operator<<(std::ostream &out, const std::vector<Term> &container)
+{
     out << "[";
     Print(out, container);
     out << "]";
@@ -41,7 +46,8 @@ std::ostream& operator << (std::ostream& out, const std::vector<Term>& container
 }
 
 template <typename Term>
-std::ostream& operator << (std::ostream& out, const std::set<Term>& container) {
+std::ostream &operator<<(std::ostream &out, const std::set<Term> &container)
+{
     out << "{";
     Print(out, container);
     out << "}";
@@ -49,7 +55,8 @@ std::ostream& operator << (std::ostream& out, const std::set<Term>& container) {
 }
 
 template <typename Key, typename Value>
-std::ostream& operator << (std::ostream& out, const std::map<Key, Value>& container) {
+std::ostream &operator<<(std::ostream &out, const std::map<Key, Value> &container)
+{
     out << "<";
     Print(out, container);
     out << ">";
@@ -57,40 +64,49 @@ std::ostream& operator << (std::ostream& out, const std::map<Key, Value>& contai
 }
 
 template <typename It>
-std::ostream& operator << (std::ostream& output, const IteratorRange<It>& pages) {
-    for (auto i = pages.begin(); i != pages.end(); i++) {
+std::ostream &operator<<(std::ostream &output, const IteratorRange<It> &pages)
+{
+    for (auto i = pages.begin(); i != pages.end(); i++)
+    {
         output << *i;
     }
     return output;
 }
 
-int main() {
+int main()
+{
     SearchServer search_server("and with"s);
-  
-    search_server.AddDocument(1, "funny pet and nasty rat"s, DocumentStatus::ACTUAL, { 7, 2, 7 });
-    search_server.AddDocument(2, "funny pet with curly hair"s, DocumentStatus::ACTUAL, { 1, 2, 3 });
-    search_server.AddDocument(3, "big cat nasty hair"s, DocumentStatus::ACTUAL, { 1, 2, 8 });
-    search_server.AddDocument(4, "big dog cat Vladislav"s, DocumentStatus::ACTUAL, { 1, 3, 2 });
-    search_server.AddDocument(5, "big dog hamster Borya"s, DocumentStatus::ACTUAL, { 1, 1, 1 });
+
+    search_server.AddDocument(1, "funny pet and nasty rat"s, DocumentStatus::ACTUAL, {7, 2, 7});
+    search_server.AddDocument(2, "funny pet with curly hair"s, DocumentStatus::ACTUAL, {1, 2, 3});
+    search_server.AddDocument(3, "big cat nasty hair"s, DocumentStatus::ACTUAL, {1, 2, 8});
+    search_server.AddDocument(4, "big dog cat Vladislav"s, DocumentStatus::ACTUAL, {1, 3, 2});
+    search_server.AddDocument(5, "big dog hamster Borya"s, DocumentStatus::ACTUAL, {1, 1, 1});
     const auto search_results = search_server.FindTopDocuments("curly dog"s);
-    int page_size = 2;
+    size_t page_size = 2;
 
     const auto pages = Paginate(search_results, page_size);
     vector<IteratorRange<vector<Document>::const_iterator>> res(pages.begin(), pages.end());
-    for (auto i : res) {
+    for (auto i : res)
+    {
         cout << i << endl;
         cout << "Page break"s << endl;
     }
     cout << "ACTUAL by default:"s << endl;
-    for (const Document& document : search_server.FindTopDocuments("funny"s)) {
+    for (const Document &document : search_server.FindTopDocuments("funny"s))
+    {
         cout << document << endl;
     }
     cout << "ACTUAL:"s << endl;
-    for (const Document& document : search_server.FindTopDocuments("big dog"s, [](int document_id, DocumentStatus status, int rating) { return status == DocumentStatus::ACTUAL; })) {
+    for (const Document &document : search_server.FindTopDocuments("big dog"s, [](int document_id, DocumentStatus status, int rating)
+                                                                   { return status == DocumentStatus::ACTUAL; }))
+    {
         cout << document << endl;
     }
     cout << "Even ids:"s << endl;
-    for (const Document& document : search_server.FindTopDocuments("big"s, [](int document_id, DocumentStatus status, int rating) { return document_id % 2 == 0; })) {
+    for (const Document &document : search_server.FindTopDocuments("big"s, [](int document_id, DocumentStatus status, int rating)
+                                                                   { return document_id % 2 == 0; }))
+    {
         cout << document << endl;
     }
     return 0;
