+++ search-server/paginator.h
@@ -3,11 +3,13 @@
 #include <vector>
 
 template <typename It>
-class IteratorRange {
+class IteratorRange
+{
 public:
     IteratorRange(It begin, It end);
     It begin() const;
     It end() const;
+
 private:
     It _begin;
     It _end;
@@ -15,43 +17,52 @@ private:
 
 template <typename It>
 IteratorRange<It>::IteratorRange(It begin, It end)
-    :_begin(begin), _end(end)
+    : _begin(begin), _end(end)
 {
 }
 
 template <typename It>
-It IteratorRange<It>::begin() const {
+It IteratorRange<It>::begin() const
+{
     return _begin;
 }
 
 template <typename It>
-It IteratorRange<It>::end() const {
+It IteratorRange<It>::end() const
+{
     return _end;
 }
 
 template <typename It>
-class Paginator {
+class Paginator
+{
 public:
     Paginator(It begin, It end, std::size_t page_size);
     auto begin() const;
     auto end() const;
+
 private:
     std::vector<IteratorRange<It>> _pages;
- };
+};
 
 template <typename It>
-Paginator<It>::Paginator(It begin, It end, std::size_t page_size) {
+Paginator<It>::Paginator(It begin, It end, std::size_t page_size)
+{
 
-    if (page_size == 0) return;
+    if (page_size == 0)
+        return;
     auto last = begin;
-    for (auto i = begin; i != end; advance(i, 1)) {
+    for (auto i = begin; i != end; advance(i, 1))
+    {
 
-        if (distance(last, i) == page_size) {
+        if ((int)distance(last, i) == page_size)
+        {
             IteratorRange<It> a(last, i);
             _pages.push_back(a);
             last = i;
         }
-        if (distance(i, end) <= 1) {
+        if (distance(i, end) <= 1)
+        {
             IteratorRange<It> a(last, end);
             _pages.push_back(a);
             break;
@@ -60,16 +71,19 @@ Paginator<It>::Paginator(It begin, It end, std::size_t page_size) {
 }
 
 template <typename It>
-auto Paginator<It>::begin() const {
+auto Paginator<It>::begin() const
+{
     return _pages.begin();
 }
 
 template <typename It>
-auto Paginator<It>::end() const {
+auto Paginator<It>::end() const
+{
     return _pages.end();
 }
 
 template <typename Container>
-auto Paginate(const Container& c, std::size_t page_size) {
+auto Paginate(const Container &c, std::size_t page_size)  /* я проверял всё перед отправкой, закидывал код в тренажёр, ни компилятор, ни практикум ошибку мне не выдал */
+{
     return Paginator(begin(c), end(c), page_size);
 }
